Implement basic logic
The program can recognize whether code editors are
running (currently Visual Studio Code (+ Insiders
version) and Pycharm).

TThen it will run the wakapi program in the
background (if it exists, of course). The script
will constantly check (with a specified delay)
whether the wakapi process is running.

to be continued...