fix(logic): :bug: Fix execution logic function
Fixed a bug where the functionality did not work at all. This was because I forgot to add the suffix to getting the “wakapi” process in the function that provides the main logic.

Renamed the “monitor_editor_toggle” function (now it is called run). This makes it clearer what it does.

The loop was also moved inside the function. It is better to call the function once than to call it many times.