feat(tray): :sparkles: Implement basic tray class
I created a “base” class for the panel (the `Tray` class). This class will be used for all interactions with the program's functionality in the panel. Currently, the class handles three functions: starting, stopping MAIN LOGIC (from the `logic.py` file), and closing the program itself.

For this, I used a library called `pystray`. It supports cross-platform functionality (but not all functions work on some platforms 😑).

Currently, this program does not have a special icon (the “standard” icon from the documentation is used), but this will be fixed in the future.